web/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-slugify==8.0.1
pydantic==2.5.3
